move store setup code to lib-store
closes #293
