+++ projects/ngx-d3timeline/src/lib/content/selectors/activity-rectangle-size.selectors.ts
@@ -1,4 +1,4 @@
-import { createSelector } from '../../selector/create-selector';
+import { createSelector } from '../../store-lib/selector/create-selector';
 import {
   selectTimeScale,
   selectBandScale
