+++ projects/ngx-d3timeline/src/lib/ngx-d3timeline.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NgxD3timelineComponent } from './ngx-d3timeline.component';
 import { ContentComponent } from './content/content.component';
 import { ActivityRectangleComponent } from './content/activity-rectangle.component';
-import { Store } from './store/store';
+import { Store } from './store-lib/store';
 import { initialState } from './store/state';
 import { of } from 'rxjs';
 import { LineComponent } from './line.component';
