+++ projects/ngx-d3timeline/src/lib/store-lib/selector/memoized-selector.ts
@@ -1,4 +1,4 @@
-import { State } from '../store/state';
+import { State } from '../../store/state';
 import { Selector } from './selector';
 import { Projector } from './types';
