+++ projects/ngx-d3timeline/src/lib/store-lib/selector/selector.ts
@@ -1,4 +1,4 @@
-import { State } from '../store/state';
+import { State } from '../../store/state';
 
 export interface Selector<TResult> {
   execute(state: State): TResult;
