+++ projects/ngx-d3timeline/src/lib/store-lib/selector/types.ts
@@ -1,4 +1,4 @@
-import { State } from '../store/state';
+import { State } from '../../store/state';
 
 export type Projector<TResult> = (...args: any[]) => TResult;
 export type SliceProjector<TResult> = (state: State) => TResult;
