+++ projects/ngx-d3timeline/src/lib/store/state.ts
@@ -1,6 +1,6 @@
 import { View } from '../view/view';
 import { TimelineDragEvent } from '../content/timeline-drag-event';
-import { createSliceSelector } from '../selector/slice-selector';
+import { createSliceSelector } from '../store-lib/selector/slice-selector';
 import { PositionedActivity } from '../positioned-activity';
 import { identifier } from '../types';
 import { defaultOptions, CompleteOptions } from '../options';
