+++ projects/ngx-d3timeline/src/lib/view/view.selectors.ts
@@ -1,6 +1,6 @@
 import { selectView } from '../store/state';
 import { getViewTopLeft, getViewClipRectangle } from './view-utils';
-import { createSelector } from '../selector/create-selector';
+import { createSelector } from '../store-lib/selector/create-selector';
 
 export const selectViewTopLeft = createSelector(selectView, getViewTopLeft);
